Split commands into a separate file, use them to drive keymaps
This adds a notion of keymaps (with mode-specific keymaps available, as well as
global keymaps) that map keys to commands.  These are the same commands as used
by the minibuffer.

There is some room to improve the flexibility of the system, but it is pretty
reasonable so far.  The keymap is stored in the state, and can therefore be
modified dynamically.