Improve the loader support
There are now loader endpoints for both LLVM and ELF.  There is also a wrapper
that tries to detect which to use.  The default for command line-provided files
is currently the file type detector.

There is also now a handle to the loading threads to allow them to be canceled
later, which will enable loading by command.