Submodule updates and loader simplification
This commit includes the update to the crucible split (into what4+crucible).

The other substantial change is a simplification of the binary loading code.  We
used to have to do a lot of pre-inspection of ELF files to set up the loader
correctly.  This was all because the loader couldn't preserve arch-specific
information for us (like the PPC TOC).

There is now a generic loader interface (currently hosted in macaw-semmc) that
preserves arch-specific information aggressively, which lets us greatly simplify
some of the loading interfaces.  Most of the complexity is now in renovate (but
can hopefully be cleaned up).