Reduce delay in the main event handling loop
Now perform a bit more computation in the worker loader thread so that the main
thread isn't hung up on some initial work before it can start handling events.

This change introduces a new index type that we can deepseq on the loader
thread, which will force evaluation of more of the code discovery.  It also
produces a useful index of a few global properties.

Note 1: It doesn't pre-compute everything, as that would take too long and
consume too much memory - most of the information is still computed on demand.

Note 2: That means we will need to be careful about querying that information,
and may have to query some things in another thread (e.g., constructing the full
CFG of a function), otherwise we'll again hang the main thread.

Note 3: This work is not yet finished.  The initial function viewer (constructed
in stateFromAnalysisResult) is still eagerly constructing the full body of the
first function in the list (it actually looks like it does that for every
analysis result it receives, and not just the first), so that will need to be
deferred to another thread and done asynchronously.

It would be best to build some generic infrastructure for doing async updates
via message passing from worker threads.  This would also be the key to doing
safer completion of filenames.