Add an experimental generic facility for asynchronous updates
The idea is that we can use this message from many different worker threads to
apply updates to the state.  This will let us more easily offload work to worker
threads to avoid blocking the event handling thread.

The goal is to use this for things like filename completion and computing CFG
properties that are too expensive to hang up the GUI, but that can't be
pre-computed during loading.