Add the ability to view the macaw version of machine code instructions
The states are not yet synced across the viewers, though the design should
support it fairly easily.  The necessary architectural changes simplified the
implementations of the function and block viewers.  There is now a simple CFG
viewer backed by an actual graph.  Note that the CFG construction is currently
incomplete, and the CFG viewer is only a stub (it is a list instead of a graph).
The new architecture is based on the notion of a stack of contexts (with a
distinguished active context) - this will allow for backwards and forwards
traversal over the context stack (the widgets already support it - a UI for
managing the stack is all that is required).

Each architecture now supports alternative IR views of each block.  For now,
only macaw is supported (only for machine code).  The same architecture supports
Crucible, too.

There is now some infrastructure for logging (with colored and configurable log levels).

This work includes a number of submodule updates.