Improve selection behavior around multiple selections and syncing
Before, the selection syncing between block viewers could select multiple
instructions at once.  The behavior of multiple selections didn't map well,
since the multiple selection is meant for persistent selections.

This commit introduces a "transient" selection that is meant for visual
highlighting, but not persistence.  If the user moves the selection, the
transient selection is replaced by a single selection, which leads to more
intuitive behavior.

The multiple selection will be used later to let users tag multiple instructions explicitly.