Refactor the minibuffer, command, and keymap to reduce the number of type parameters
There used to be a large number of related type parameters, which led to
excessively-long type signatures.  This change modifies these types to be based
on a new 'CommandLike' typeclass that exposes a set of type families instead.
This greatly simplifies most type signatures (at the expense of requiring a few
complicated kind signatures in a few places).

As an additional benefit, this change will make it easier to re-use this
functionality outside of surveyor, as the core of keymap/command/minibuffer are
even less closely coupled to the peculiar needs of surveyor.