Add filename completion
It is a bit quirky right now - it is closely tied to the semantics of the Glob
package.  Something more intuitive might be worthwhile...

Glob is very sensitive to path separators and extensions.  Getting what you want
is easy, but it might be better to pursue something easier to reason about.