Remove a redundant nonce from the arch state
The arch nonce is now floated to the top-level of the state so that we always
have one, even if there is no arch state (in that case, the arch parameter is
Void).

Since we always have one, we can clean up some code that dealt with cases where
the arch nonce was missing.