Change the state type for SurveyorCommand
The state type in commands is the type of a "state" value passed to each command
handler callback.  The old state type was just the arch nonce, which was enough
for most things.  In fact, it was enough for all current use cases.

The next step is to add some functionality to commands to let them decide if
they can be applied in the current context (i.e., state).  That function will
basically be of the form `StateType x -> Bool`.  This required the
generalization.  Unfortunately, it had some knock-on effects that touched a few
type signatures and required yet another wrapper around State to quantify out
some type parameters (because there are many places where we can't mention arch).