Allow commands to determine if they can be run based on the current state
This is now used in the block viewer rendering to only show commands that can be
run given the current context.  Currently, all commands can always be run.  That
won't always be true, though (e.g., a command to follow a branch can only run if
a branch is selected).