Implement a more structured operand list for the IR
Previously, operands were just stored in a plain Haskell list.  This was a bit problematic, because
it made it difficult to group delimited (sub-)sequences of related operands in a reasonably
structured way.  One way is to introduce "pseudo" operands to represent syntax, but this makes it
hard to get the comma separation right (since you would start putting commas between syntax).

Now, operands are in a structured OperandList, with an explicit representation of nesting and
delimiter syntax.  All of the contents are semantically-meaningful operands.  We now track which
operand is focused through a zipper.  Note that we also have to explicitly number each
operand (accomplished via Traversable) so that we can give operands identity; this matters, since
the same operand can appear multiple times in an operand list, and so could compare equal to the
selected operand, even if a different instance was selected.  We track identity with simple
integers.

As a side effect of switching to zippers, some of the ugly logic around selected operand index
manipulation has been removed.

This change will enable us to more easily break down large compound operands into something
reasonable (e.g., register states in macaw).