Modify asAlternativeIR to preserve more information
It still returns the (possibly incomplete) block mapping, but now additionally
returns all of the alternative IR blocks (even ones with no obvious mapping to
base blocks).

This is still useful for navigating the alternative view of the function, just
without the ability to necessarily switch.  This will be important for building
out the symbolic execution UI