Separate symbolic execution state from the context stack
Updates to the symbolic execution state can arrive asynchronously (and apply to
not just the top of the context stack).  Inserting these updates into the
context stack would be both tricky and expensive.  This commit changes things
around so that we can store symbolic execution state outside of the context
stack, which is much easier to update.