ghc 8.8 compat
A more precise (dependent) kind signature was required for commands.  8.8 is
more strict (and probably more correct), noting that the old signature was
ambiguous (there was a heterogeneous type equality error).

Note that ArgumentKind had to be moved from an associated type to a standalone
type family because recursive groups of associated type definitions are not allowed.