Feature/symbolic execution (#1)
Add support for symbolic execution inside of surveyor

All of the backends are at least stubbed out and compiling, but the llvm one is actually fully implemented.  It provides a UI for selecting a function and symbolically executing it.  It streams intermediate results into the application state, which can be monitored.

Changes rolled up into this:
* Crucible CFG construction for all backends
* Separation of the `OperandList` into its own module with better abstraction boundaries
* GHC 8.8 compat