Use lumberjack as the logging core (#3)
This regularizes the logging infrastructure to not emit messages and use the
logging API directly.  This will let us more easily support sending logs to
files as well as the in-memory buffer.

This uses a custom log message type modeled after the one from lumberjack.  It
attempts to hide the fact that it uses lumberjack to make future updates to call
sites a bit easier (hopefully).

Also adds a mechanism for logging to files (in addition to the in-UI log viewer)