WIP Start implementing a Crucible execution feature
This creates an initial state that currently always interrupts Crucible on every
step to spawn Surveyor.

That is obviously too aggressive, but it is an experiment in making sure that it
is possible to implement all of the necessary data structures.

It looks like we can, largely.  There are still some interesting questions to
answer about deconstructing the suspended state to populate e.g., the current
CFG.

We also need to be more discriminating on when we trigger it.  We could trigger
on any call to the breakpoint construct.  That would make sense if the
breakpoint had no arguments.  If there was a conditional breakpoint, we would
need to actually evaluate it.  That might be better implemented via an
override (but most of the same code would still be useful for converting a
"live" symbolic execution state into surveyor).

The execution feature could even be made programmable with a predicate over
Crucible states that indicated when to fire.