Add a modified version of crux-llvm, crux-dbg, with debugging support
This also adds a `crucible_breakpoint` primitive to crucible.h.  When crux-dbg
encounters a `crucible_breakpoint`, it brings up the brick UI for surveyor.
Currently, the UI view isn't very useful.  However, it will allow us to inspect
the current symbolic state (and potentially modify it and continue execution).

The tool also supports a `--interactive` flag, which immediately brings up the
TUI instead of starting crux.

The idea is that this tool could stand in for all of the crux-* tools, while
bringing TUI/debugging support.