Ensure we display the symbolic execution manager when entering from crux-dbg
There was some logic that always showed the intro text if no filename is set.
That is overly-aggressive.

NOTE: We could also plumb a filename through the execution feature, but that is
really just a bad design in surveyor.  There is no guarantee that there will
ever be a unique filename to work from.