Add a new tool: crux-dbg and a debugging execution feature
This change adds two things: crux-dbg and a debugging execution feature for Crucible

- The execution feature allows a Crucible-based tool to drop into surveyor-brick as a debugger
- `crux-dbg` is a new crux frontend that essentially extends `crux-llvm` to use the debugger execution feature

The new `crux-dbg` tool currently accepts only LLVM bitcode as an input. `crux-llvm` also accepts C/C++ files and compiles them to bitcode automatically. `crux-dbg` reduces its implementation complexity by eliding this feature.  It otherwise supports everything that `crux-llvm` supports, particularly with respect to overrides.  It introduces a new override: `crucible_breakpoint`.   `crux-dbg` is set up to drop into the surveyor-brick TUI when symbolic execution reaches a `crucible_breakpoint`.  There is a new UI in surveyor to support inspecting a suspended symbolic execution state (which is available while crucible is paused). Once the TUI is exited, control is returned to crux, which will continue verification.

The new UI is primitive, but is ripe for development of richer debugging workflows.

`crux-dbg` is intended to become a multi-language verifier.  It also supports a `--interactive` flag to bring up the surveyor-brick TUI instead of starting crux.

Fixes #6