Progress on the RegValue viewer
It looks like this approach will work and it is mostly a matter of filling out
cases now in the rendering code.

There is one issue that needs to be resolved upstream.  In order to traverse
symbolic terms, we need to know that `sym ~ ExprBuilder`; however, this
constraint is not available from the `ExecutionFeature` context, where `sym` is
totally abstract.  We will need to concretize the type in `SimulatorCallback`.