Tr/breakpoint names (#54)
Add the infrastructure to render breakpoint names

It was tricky because reading strings in the LLVM backend requires a access to the memory model and IO.  This required changes to crux.

This also adds a widget for viewing the call stack at a breakpoint. Right now, it lists all of the override, call, and return frames, but doesn't let you do much with them.  For call frames, it shows the basic block corresponding to the call.

In the longer term, we can use this UI to inspect values in the call stack (from the full state viewer).  Each call frame has the same type of register list as breakpoints, so it should be straightforward.