Add a function to force-invoke a command through the minibuffer
This invokes a command and allows the minibuffer to prompt the user for
arguments interactively.  It isn't particularly magical, but the necessary
incantation is also not obvious.

It may make sense to use this function to implement the same logic in the
minibuffer core, but I think ultimately that code might need some different
logic.