Add infrastructure for letting users specify names for values
Add the ability for users to name sub-terms in the `ValueViewer` widget. Users can select sub-terms and use a new contextual command to provide a name.  The TUI prompts the user via the minibuffer for the requested name.  The `ValueViewer` has been updated to render names next to relevant terms.

Currently, only terms with nonces can be named.

Fixes #59.  Begins addressing #57.