Fix the symbolic execution configuration form
The observed issue was that selecting a solver and solving mode in the
initialization UI was not reflected in the setup UI on the next page when the
BeginSymbolicExecution command was invoked.

The cause was that the symbolic execution session ID was stale.  I didn't look
closely, but assume that we assigned an initial state there at some point.  In
the handler for `InitializeSymbolicExecution`, we created a fresh state (via
`configuringSymbolicExecution` and either the `mConfig` that was passed in or
the call to `defaultSymbolicExecutionConfig` with a fresh nonce).  This new
config has its own session ID that we inserted into the map containing all of
the active sessions, but we never set the active sessionID to that value.

The fix was in that handler, where we now modify the session ID in the current
context.