Auto-assign names to function arguments in the state viewer
Users can manually name values, but that requires work.  This change
automatically assigns names to values for which we can come up with sensible
names: namely function arguments.

This requires a bit of wrangling because most LLVM arguments are of type
LLVMPointer, which does not have nonces associated with it.  This change also
introduces a method in the SymbolicArchitecture class to let us assign names to
intrinsic typed arguments.  In particular, we assign individual names to the
BlockID and Offset in each LLVMPointer.

Note, morally this works but in practice I think we have a slight problem: our
translation of LLVM makes a synthetic entry block with all of the formal
parameters to the function that immediately jumps to another block.  This means
that the entry block doesn't seem to actually be in scope in the call stack,
which means that we don't get names for function arguments.  At this point, I
don't believe that we can really recover the original function arguments, as the
block is inaccessible.