Redo the symbolic execution state stepping code
This change uses a completely different method than before.  The previous
implementation attempted to manually resume execution by returning out of
overrides to the caller.  This was technically possible, but made further
stepping difficult (since it effectively required recursive crucible
invocations).

The new model inserts a simple execution feature that enables single stepping at
a very granular level, while not interfering with any of the normal
branching/merging logic (it inherits whatever the top-level Crucible symbolic
execution strategy is).

It runs surveyor and crux in separate threads and uses channels to communicate
between the two.  The crux thread can stop due to the debugging execution
feature or due to debugging overrides (e.g., breakpoint).  When crux stops, it
sends the current simulator state to surveyor over the channel, allowing
surveyor to set up suitable execution frames for interactive inspection.
Stepping or continuing becomes just sending a signal (and possibly a modified
simulator state) back to crux and resuming it.

The mode of the execution feature is governed by a global value in an IORef.

The major complexity is in the setup of the simulator and surveyor, as well as
the protocol that governs their interactions.  The protocol isn't yet entirely
correct (and can hang if the user doesn't explicitly continue before exiting
surveyor).

This new formulation also makes some more advanced features like stepping
out/over possible.