Add an interrupt command
This interrupts a running symbolic execution.  Note that it currently needs the
execution to be running with the UI enabled.  It could be invoked by e.g., an
interrupt handler that could let the user start the UI when a SIGUSR1 is
received.

This change also generalizes the debugging events to carry a SessionID to
operate on, rather than just always operating on the current session.  The
commands that generate these events now take the current session.  This should
be more predictable and enable more workflows based on events in the future.