Redo the symbolic execution state stepping code (#78)
This change uses a completely different method than before.  The previous
implementation attempted to manually resume execution by returning out of
overrides to the caller.  This was technically possible, but made further
stepping difficult (since it effectively required recursive crucible
invocations).

The new model inserts a simple execution feature that enables single stepping at
a very granular level, while not interfering with any of the normal
branching/merging logic (it inherits whatever the top-level Crucible symbolic
execution strategy is).

It runs surveyor and crux in separate threads and uses channels to communicate
between the two.  The crux thread can stop due to the debugging execution
feature or due to debugging overrides (e.g., breakpoint).  When crux stops, it
sends the current simulator state to surveyor over the channel, allowing
surveyor to set up suitable execution frames for interactive inspection.
Stepping or continuing becomes just sending a signal (and possibly a modified
simulator state) back to crux and resuming it.

The mode of the execution feature is governed by a global value in an IORef.

The major complexity is in the setup of the simulator and surveyor, as well as
the protocol that governs their interactions.  The protocol isn't yet entirely
correct (and can hang if the user doesn't explicitly continue before exiting
surveyor).

This new formulation also makes some more advanced features like stepping
out/over possible.

* Fix a deadlock on cleanup

Previously, if surveyor was shut down while a symbolic execution instance was
suspended, the whole system would deadlock (and not even respond to Ctrl+C).
Surveyor now explicitly turns off its hooks into the symbolic execution engine
and resumes all outstanding executions when shutting down.

This behavior works well for crux-dbg, but might not be appropriate for jobs
started by surveyor itself.  A distinction will probably need to be made later
to distinguish the two.

* Add an interrupt command

This interrupts a running symbolic execution.  Note that it currently needs the
execution to be running with the UI enabled.  It could be invoked by e.g., an
interrupt handler that could let the user start the UI when a SIGUSR1 is
received.

This change also generalizes the debugging events to carry a SessionID to
operate on, rather than just always operating on the current session.  The
commands that generate these events now take the current session.  This should
be more predictable and enable more workflows based on events in the future.

* Allow users to interrupt crux-dbg with SIGUSR2

On UNIX platforms, allow users to trigger the debugger UI by sending SIGUSR2 to
crux-dbg.  This will pause the symbolic execution at the next Crucible hook
point and bring up the surveyor UI with a suspended symbolic execution session.

Note: on Windows this functionality is not supported (but crux-dbg still works on Windows otherwise)

* Rewrite the communication of suspend events

The new approach makes all of the reasons for suspended execution explicit,
removes some duplicate state information, and makes rendering suspended states
simpler.