Render basic blocks in the symbolic call stack viewer (#79)
* Revise the block renderer API

Instead of rendering based on the current context, render just the block (state)
that is passed in.  This will make it significantly easier to reuse in new
contexts (where we might not want to modify the context).

* Slightly improve the rendering of the current call stack

* Render Crucible blocks in the (symbolic) call stack viewer

This re-uses the basic block renderer to show the Crucible basic block for each
frame on the stack.  This required extracting and generalizing some of the
original Crucible block translation code.  Right now, it is only implemented for
LLVM (machine code and JVM leave this safely unimplemented).