Implement a few commands supporting *replay debugging*
This introduces a command to start recording traces of symbolic execution
states.  There are additional commands to step backward and forward through the
trace.  This is much like replay debugging in e.g., gdb, but over symbolic
states.

Note that this is a raw back/forward trace stepping now, in the order that
Crucible visited states.  There is another axis (across symbolic branches) that
might be more interesting.  That could be implemented on top of this raw trace.