Move initial symbolic registers
When setting up a new symbolic execution session from within the surveyor
core (instead of outside via e.g. crux), store the initial registers in the
symbolic execution mode constructor, rather than the configuration state.  It
isn't used in all symbolic execution tasks, and is not available if we are
connecting to a session initiated by crux.

Fixes #83