Make the minibuffer respect predicates indicating command applicability
Each command has a predicate over the current state that says whether or not it
can be used.  Running a command that is not currently applicable is not a
problem, as it is just a no-op, but it can be confusing to users when nothing
happens.

This change hides commands from the completion list if they cannot currently be
used.

Note that keybindings already respected these predicates.

Fixes #61