Revise value viewer cache handling
Before, only values with nonces were cached.  This meant that rendering around
aggregate Crucible values (e.g., LLVMPointer) was a bit weird and did not
participate in the caching.  This led to some rendering issues.

This change adds a secondary cache (based on surrogate keys) that is able to
cache values without nonces.