Add a new abstraction for implementing event handlers
This is an alternative to the extensive use of pattern guards in the existing
handlers.  This formulation allows for more granular error reporting in the case
that a handler cannot execute.  Currently, all of the pattern guards either
succeed or fail.  However, some would benefit from additional diagnostic output
to make failure more apparent (either via the log or minibuffer echo area).

The rest of the handlers should be updated when there is time.