Move the EchoArea into the Brick UI
This didn't really need to be in the core, except that a few core functions
explicitly generated echo area updates.

This commit introduces a new policy that just reflects any non-debug log
message (at least the first line) into the echo area.  This means that the brick
UI can just postprocess any log events and update the UI if needed.  The result
is a slight simplification and is logically a bit nicer.

Fixes #92